Minikube env var for deploy/deletion scripts.

+++ scripts/delete-debug-ds.sh
@@ -1,4 +1,9 @@
 #!/usr/bin/env bash
 
+if [[ -n "${TNF_MINIKUBE_CLUSTER}" ]]; then
+    echo "minikube detected, skip deleting debug daemonSet"
+    exit
+fi
+
 oc delete daemonsets.apps/debug -n default
 until [[ -z "$(oc get ds debug -n default 2>/dev/null)" ]]; do sleep 5; done
